+++ routes/api/review.js
@@ -7,9 +7,14 @@ router = express.Router();
 router.get("/",async (req,res)=>{
     /// get a list of all reviews in the database
     try {
+        if(req.isAuthenticated()){
+
             const review = await Review.find()
             res.json(review);
-
+        }
+        else{
+            res.status(401).send("Unauthorized")
+        }	
     }
     catch (error) {
         res.json({message: error})
@@ -20,7 +25,7 @@ router.get("/",async (req,res)=>{
 router.post("/add",async (req,res)=>{
     try {
         
-        // if(req.isAuthenticated()){
+        if(req.isAuthenticated()){
             const review_object = new Review({
                 Parent: req.body.Parent,
                 Author: req.body.Author,
@@ -33,10 +38,10 @@ router.post("/add",async (req,res)=>{
             const savedReview = await review_object.save();
             console.log("New Review added successfully");
             res.json({status:true, newReview: review_object});
-        // }
-        // else{
-        //     res.status(401).send("Unauthorized")
-        // }	
+        }
+        else{
+            res.status(401).send("Unauthorized")
+        }	
     } catch (error) {
         res.json({message: error});
     }
@@ -45,13 +50,13 @@ router.post("/add",async (req,res)=>{
 
 router.get('/:id',async (req,res)=>{
     try {
-        // if(req.isAuthenticated()){
+        if(req.isAuthenticated()){
             const review = await Review.findById(req.params.id);
             res.json(review); 
-        // }
-        // else{
-        //     res.status(401).send("Unauthorized")
-        // }
+        }
+        else{
+            res.status(401).send("Unauthorized")
+        }
     } catch (error) {
         res.json({message: error});
     }
@@ -60,15 +65,15 @@ router.get('/:id',async (req,res)=>{
 router.patch("/:id",async (req,res)=>{
     /// update the attributes of a course in the database 
     try {
-        // if(req.isAuthenticated()){
+        if(req.isAuthenticated()){
             const updatedReview = await Review.updateOne({_id: req.params.id},  {$set: req.body});
             const review = await Review.findById(req.params.id);
             console.log("Professor updated successfully");
             res.json(review);
-        // }
-        // else{
-        //     res.status(401).send("Unauthorized");
-        // }
+        }
+        else{
+            res.status(401).send("Unauthorized");
+        }
     } catch (error) {
         res.json({message: error});
     }
@@ -77,13 +82,13 @@ router.patch("/:id",async (req,res)=>{
 
 router.delete('/:id',async (req,res) => {
     try {
-        // if (req.isAuthenticated()) {
+        if (req.isAuthenticated()) {
             const removedReview = await Review.deleteOne({_id: req.params.id});
             console.log("Review deleted successfully")
             res.json(removedReview)
-        // } else {
-        //     res.status(401).send("Unauthorized");
-        // }
+        } else {
+            res.status(401).send("Unauthorized");
+        }
     } catch (error) {
         res.json({message: error})
     }
